requirements.txt
git+https://github.com/Pycord-Development/pycord
python-dotenv
requests
deta